test: more reliable tests using new 'test mode'
