+++ test/components/RenderCacheable/component.nuxt.spec.ts
@@ -56,6 +56,7 @@ vi.mock('#nuxt-multi-cache/config', () => {
     debug: false,
     cdnEnabled: true,
     routeCacheEnabled: true,
+    isTestMode: false,
   }
 })
