+++ test/composables/useCDNHeaders.nuxt.spec.ts
@@ -43,6 +43,7 @@ vi.mock('#nuxt-multi-cache/config', () => {
     cdnEnabled: true,
     cdnCacheControlHeader: 'Surrogate-Control',
     cdnCacheTagHeader: 'Cache-Tag',
+    isTestMode: true,
   }
 })
