+++ test/composables/useDataCacheCallback.nuxt.spec.ts
@@ -1,10 +1,8 @@
-import type { H3Event } from 'h3'
 import { describe, expect, test, vi, afterEach, beforeEach } from 'vitest'
 import { useDataCacheCallback } from './../../src/runtime/composables/useDataCacheCallback'
 import { useDataCache } from './../../src/runtime/composables/useDataCache'
-import type { CacheItem } from './../../src/runtime/types'
-import { MULTI_CACHE_CONTEXT_KEY } from '~/src/runtime/helpers/server'
 import { toTimestamp } from '~/src/runtime/helpers/maxAge'
+import { buildEvent, buildEventWithStorage } from '../__helpers__/event'
 
 let isServerValue = false
 
@@ -16,67 +14,9 @@ vi.mock('#nuxt-multi-cache/config', () => {
     debug: false,
     routeCacheEnabled: false,
     cdnEnabled: false,
+    isTestMode: true,
   }
 })
-
-function buildEventWithStorage(storage: Record<string, CacheItem>): H3Event {
-  return {
-    context: {
-      [MULTI_CACHE_CONTEXT_KEY]: {
-        cache: {
-          data: {
-            storage: {
-              getItem: (key: string) => {
-                if (key === 'force_get_error') {
-                  throw new Error('Failed to get data cache item.')
-                }
-                return Promise.resolve(storage[key])
-              },
-              setItem: (key: string, data: any) => {
-                storage[key] = data
-                return Promise.resolve()
-              },
-            },
-          },
-        },
-      },
-    },
-  } as H3Event
-}
-
-function buildEvent(bubbleError = false): H3Event {
-  const storage: Record<string, CacheItem> = {
-    foobar: { data: 'Cached data.', expires: -1, staleIfErrorExpires: 0 },
-    expires: {
-      data: 'Data with expiration date.',
-      expires: 1669849200,
-      staleIfErrorExpires: 0,
-    },
-  }
-  return {
-    context: {
-      [MULTI_CACHE_CONTEXT_KEY]: {
-        cache: {
-          data: {
-            bubbleError,
-            storage: {
-              getItem: (key: string) => {
-                if (key === 'force_get_error') {
-                  throw new Error('Failed to get data cache item.')
-                }
-                return Promise.resolve(storage[key])
-              },
-              setItem: (key: string, data: any) => {
-                storage[key] = data
-                return Promise.resolve()
-              },
-            },
-          },
-        },
-      },
-    },
-  } as H3Event
-}
 
 vi.mock('#imports', () => {
   return {
