+++ test/fileSystemDriver.e2e.spec.ts
@@ -13,6 +13,7 @@ vi.mock('#nuxt-multi-cache/config', () => {
       return true
     },
     debug: false,
+    isTestMode: true,
   }
 })
